Ajustes ordenes y clientes

+++ services/Clientes/APIClientes/src/main/java/com/touresbalon/api/repository/read/ReadClienteEntity.java
@@ -19,7 +19,7 @@ public class ReadClienteEntity {
 
 
 	public ReadClienteEntity(String identificacion, String nombres, String apellidos, String telefono, String email,
-			String celular, @Valid @DecimalMin("1") int estado, @Valid @DecimalMin("1") BigDecimal categoria,
+			String celular, @Valid @DecimalMin("1") int estado, @Valid @DecimalMin("1") Integer categoria,
 			String tipoIdentificacion, Timestamp fechaRegistro) {
 		this.identificacion=identificacion;
 		this.nombres=nombres;
@@ -63,7 +63,7 @@ public class ReadClienteEntity {
 	private int estado;
 
 	@Column(name="categoria")
-	private BigDecimal categoria;
+	private Integer categoria;
 
 	@Column(name="tipoidentificacion")
 	private String tipoIdentificacion;
